// src/app/shared/interfaces/Tiempo/tiempo-pais-interface.ts
export interface TiempoPais {
    name?: string,
    tem?: number,
    temp_min?:   number;
    temp_max?:   number;
    humidity?:   number;
    pressure?:   number;
    icon?:       string;
    wind_speed?: number;
    visibility?: number;
    timezone?:   number;
    weatherdescripcion?: string;

}